requirements.txt
azure_storage==0.37.0
duckdb==1.3.2
langchain_core==0.3.75
langchain_openai==0.3.32
langgraph==0.6.6
mermaid==0.3.2
nest_asyncio==1.6.0
pandas==2.3.2
Pillow==11.3.0
pydantic==2.11.7
pydantic_settings==2.10.1
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
streamlit==1.48.1
